requirements.txt
# SensorGuard Python dependencies
# Install with: pip install -r requirements.txt
psutil>=5.9
websockets>=12.0
watchdog>=3.0
wmi>=1.5.1
pywin32>=306
openai>=1.0
python-dotenv>=1.0
